extract signal, file modes and inode support checks to CapabilityCheckHelpers for reuse
