+++ spec/spec_helper.rb
@@ -73,3 +73,23 @@ RSpec::Matchers.define :be_similar_to do |expected, max_difference|
         "#{expected}, got normalized root-mean-square error of #{@diff}"
   end
 end
+
+module CapabilityCheckHelpers
+  def any_file_modes_allowed?
+    Tempfile.open 'posix' do |f|
+      File.chmod(0, f.path)
+      File.stat(f.path).mode & 0o777 == 0
+    end
+  end
+
+  def inodes_supported?
+    File.stat(__FILE__).ino != 0
+  end
+
+  def signals_supported?
+    Process.kill(0, 0)
+    true
+  rescue
+    false
+  end
+end
